+++ src/components/CategoryFilter/CategoryFilter.scss
